Add files via upload
January 29, 2022: White paper ver 5-1-0 is available. This update provides the Generation 12 vaccine design.

CRITICAL UPDATE: One of the peptides designed to represent the omicron receptor binding motif (RBMO-sc) in Generation 12 has repeatedly failed chemical synthesis. Note that without this peptide, Generation 12 is not specific to omicron, and only provides conserved epitopes largely shared with Generation 11. If you decide to attempt to synthesize this peptide (or a modified sequence), please contact us at info@radvac.org. Due to this failure and recognition of the possibility that peptide synthesis might fail at a key moment of future deployment, we are working on an omicron-specific vaccine design based on RNA, which we will release shortly.